requirements.txt
python-telegram-bot==13.7
PyPDF2
ebooklib
Pillow
Flask
reportlab